Big update in PDF generated.
Added a detailed PDF comparing radial upstream and parallel to s OEDGE to Thomson comparisons.

GUI updated accordingly, but still needs options added to it to take full advantage of the PDF options.

Also added ability to generate Thomson Excel files mapped to a common flux surface into oedge_plots, as well as a helper function to make sure they are plotted in the right area.